do light validation on import tool, switch from NSUInteger to its prettier sibling unsigned int

+++ TFMMobileMarketManager/ImportTool/ImportTool.h
@@ -17,8 +17,8 @@
 
 - (id)initWithSkipSetting:(bool)skipFirstRow;
 
-- (NSUInteger)importStaffFromCSV:(NSURL *)url;
-- (NSUInteger)importVendorsFromCSV:(NSURL *)url;
-- (NSUInteger)importLocationsFromCSV:(NSURL *)url;
+- (unsigned int)importStaffFromCSV:(NSURL *)url;
+- (unsigned int)importVendorsFromCSV:(NSURL *)url;
+- (unsigned int)importLocationsFromCSV:(NSURL *)url;
 
 @end
